+++ src/__init__.py
@@ -1,1 +1,3 @@
 from config import load_config
+
+from work_queue import TTWorkQueue