+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.0.3
+* Added reconnect callback to TTWorkQueue
+
+## v0.0.2
+
+* Added TTWorkQueue class
+
 ## v0.0.1:
 
 * Initial release